requirements.txt
flask==3.0.2
openai==1.11.0
feedparser==6.0.11
PyGithub==2.1.1
python-dotenv==1.0.1
Jinja2==3.1.3
pytest==8.0.2
requests==2.31.0 